auditoria(A5): regeneracion de tests/integration con el nuevo ABI (int64_t/double) — _cripto.c/_sistema.c/_synapse_shared.h/_test_cluster_handshake.c: ed25519_verificar/ejecutar_comando -> int64_t, NODO_NULO..CONSTRUCTOR (47-52) en el header compartido

+++ tests/integration/_cripto.c
@@ -1,6 +1,6 @@
 #include "_synapse_shared.h"
 
-int ed25519_verificar(CadenaSegura mensaje, CadenaSegura firma, CadenaSegura clave_publica) {
+int64_t ed25519_verificar(CadenaSegura mensaje, CadenaSegura firma, CadenaSegura clave_publica) {
     return _syn_ed25519_verificar(mensaje, firma, clave_publica);
       /* [Lifetime Scope: exit depth=0] */
 }
